Update views.md
The view dropdown is the rightmost menu of those that are located on the top left. This makes it confusing to describe but I think this revision could help clarify it at least some. The original statement had me scouring the top of the map trying to find a view menu on the right side which is entirely the wrong spot to look (and as such I thought something was going wrong): "Use the navigation at the top of your map and click the view dropdown menu on the right."

Nothing critical at all, it is just slightly clunky enough I could see someone getting confused for a bit. Even just further clarification that navigation is specifically those menus on the left would fix it. "Navigation at the top of your map" made me think that whole area counted so I was looking for extra menus I thought weren't appearing. It makes sense once you know what it means but was confusing to understand before that.